# Trash/consolidation_2026-07-23_P2/src/bridges/roma/openapi.yaml
# Deprecated: use canonical spec at docs/api/openapi.yaml
# This file is kept only for archaeology; see PR audit/doc04-openapi.

# ROMA Platform — OpenAPI Specification
# Standardizes ROMA as API-first execution kernel

openapi: 3.1.0
info:
  title: ROMA (Role-Oriented Multi-Agent Execution Platform)
  version: 1.0.0
  description: |
    Deterministic execution kernel for GPU workloads on Kubernetes + Ray.
    NOT a generative AI — strict input contract, no fallback mode.
  contact:
    name: asurdev
    url: https://asurdev.zo.computer

servers:
  - url: http://localhost:8000
    description: ROMA Execution Bridge

paths:
  /submit:
    post:
      operationId: submitTask
      summary: Submit ROMA task (strict validation, no fallback)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RomaTaskInput'
      responses:
        '202':
          description: Task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RomaTaskResponse'
        '400':
          description: Validation error (input rejected)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RomaRejectResponse'
        '429':
          description: Backpressure (queue saturation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RomaRejectResponse'
        '500':
          description: Internal execution error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RomaRejectResponse'

  /status/{job_id}:
    get:
      operationId: getTaskStatus
      summary: Get task execution status
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RomaStatusResponse'
        '404':
          description: Task not found

  /queue/status:
    get:
      operationId: getQueueStatus
      summary: Get queue metrics (depth, GPU saturation)
      responses:
        '200':
          description: Queue status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RomaQueueStatus'

  /health:
    get:
      operationId: healthCheck
      summary: ROMA health check
      responses:
        '200':
          description: System healthy

  /metrics:
    get:
      operationId: getMetrics
      summary: Prometheus-compatible metrics
      responses:
        '200':
          description: Metrics in text format

components:
  schemas:
    RomaTaskInput:
      type: object
      required: [task]
      properties:
        task:
          type: string
          description: User task description (MANDATORY, non-empty)
          minLength: 1
          example: "train YOLOv8 on RTX3060 with COCO dataset"
        priority:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
        gpu_required:
          type: boolean
          default: false
        execution_mode:
          type: string
          enum: [k8s_job, ray_job]
          default: k8s_job
      # STRICT: additional properties not allowed
      additionalProperties: false

    RomaRejectResponse:
      type: object
      properties:
        status:
          type: string
          const: "rejected"
        error:
          type: object
          properties:
            code:
              type: string
              enum: [USER_TASK_REQUIRED, VALIDATION_ERROR, BACKPRESSURE, INTERNAL_ERROR]
            message:
              type: string
            severity:
              type: string
              enum: [critical, high, medium]
          required: [code, message, severity]
        roma_validation:
          type: object
          properties:
            input_received:
              type: [string, "null"]
            input_length:
              type: integer
            validation:
              type: string
              const: "STRICT"
            fallback_mode:
              type: boolean
              const: false
        behavior:
          type: string
          const: "exit_immediately"

    RomaTaskResponse:
      type: object
      properties:
        status:
          type: string
          const: "queued"
        job_id:
          type: string
          format: uuid
        roma_dispatch:
          type: object
          properties:
            submit_target:
              type: string
              const: "execution_bridge"
            queue_priority:
              type: integer
              minimum: 1
              maximum: 10
        dag:
          type: array
          items:
            $ref: '#/components/schemas/RomaTaskStep'
        estimated_resources:
          type: object
          properties:
            cpu:
              type: string
            ram:
              type: string
            vram:
              type: string
        gpu_required:
          type: boolean

    RomaTaskStep:
      type: object
      properties:
        step:
          type: string
        depends_on:
          type: array
          items:
            type: string
        action:
          type: string
        image:
          type: string
        gpu:
          type: boolean
      required: [step, action]

    RomaStatusResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, scheduled, running, completed, failed, cancelled]
        created_at:
          type: string
          format: date-time
        started_at:
          type: [string, "null"]
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
        k8s_job_name:
          type: [string, "null"]
        node_name:
          type: [string, "null"]
        error:
          type: [string, "null"]

    RomaQueueStatus:
      type: object
      properties:
        depth:
          type: integer
        gpu_available:
          type: integer
        gpu_used:
          type: integer
        saturation:
          type: number
          format: float
        backpressure_active:
          type: boolean